Move code required by libufs into libufs
Needed to move more code into sys/include.  Also split some functions out of ffs_subr into libufs/libufs_subr.c
