+++ sys/src/9/amd64/core.json
@@ -31,7 +31,8 @@
 			"/$ARCH/lib/klibc.a",
 			"/$ARCH/lib/klibdraw.a",
 			"/$ARCH/lib/klibip.a",
-			"/$ARCH/lib/klibsec.a"
+			"/$ARCH/lib/klibsec.a",
+			"/$ARCH/lib/klibufs.a"
 		],
 		"Oflags": [
 			"-z",
