+++ sys/src/9/riscv/core.json
@@ -20,7 +20,8 @@
 		"Libs": [
 			"/$ARCH/lib/klibc.a",
 			"/$ARCH/lib/klibip.a",
-			"/$ARCH/lib/klibsec.a"
+			"/$ARCH/lib/klibsec.a",
+			"/$ARCH/lib/klibufs.a"
 		],
 		"Oflags": [
 			"-z",
