+++ sys/src/libufs/ffs_tables.c
@@ -31,8 +31,9 @@
 
 #include <u.h>
 #include <libc.h>
-#include <ufs/dinode.h>
-#include <ffs/fs.h>
+
+#include "ufs/dinode.h"
+#include "ffs/fs.h"
 
 /*
  * Bit patterns for identifying fragments in the block map
